Improve the sheet motion while opening/closing the keyboard (#27)
Fixes #22. Unfortunately, there are still some problems related to the
keyboard (e.g. #14) that will be taken care of in a future PR.

Summary:

- Added `SheetExtent.settle`, which causes the sheet to settle into a
stationary position similar to the `goBallistic(0.0)`, but with a
constant (non zero) and relatively high velocity.
- Added `SheetActivity.didFinalizeDimensions`, which is called once in a
frame after all measurement processes for both the content and the
viewport size have been completed.