Fix: Opening keyboard interrupts sheet animation (#189)
Fixes #14.

The sheet position adjustment strategy of `AnimatedSheetActivity`, which
is used when either the sheet content size or the viewport size changes,
was modified to:

1. append the delta of `MediaQueryData.viewInsets.bottom` (the keyboard
height) to keep the visual sheet position unchanged, and
3. if the animation is still running, start a new linear animation to
bring the sheet position to the recalculated final position in the
remaining duration. We use a linear curve here because starting a curved
animation in the middle of another curved animation tends to look jerky.