Prefer composition style for SheetKeyboardDismissible (#197)
## Fixes / Closes (optional)

None.

## Description

This PR prepares for the upcoming `SheetActivity.ignorePointer` API. In
these changes, the implementation of `SheetContext` will be moved from
`SheetExtentScopeState` to the State classes of each sheet widget. As a
result, the `SheetKeyboardDismissible` widget will need to be an
ancestor of a sheet. This is because `SheetContext.notificationContext`
will be at the sheet widget itself (currently it is at
`SheetExtentScope`, which is a descendant of the sheet). Consequently,
the `SheetKeyboardDismissible` internally inserted into the sheet
subtree won't be able to catch `SheetNotification`s, resulting in the
keyboard dismissing behavior not working.

## Summary (check all that apply)

- [x] Modified / added code
- [x] Modified / added tests
- [x] Modified / added examples
- [ ] Modified / added others (pubspec.yaml, workflows, etc...)
- [ ] Updated README
- [x] Contains breaking changes
  - [x] Created / updated migration guide
- [ ] Incremented version number
  - [ ] Updated CHANGELOG