Decrease required minimum pub points (tmp fix) (#216)
## Fixes / Closes (optional)

None.

## Description

After the release of Flutter 3.24, pub.dev and `pana` have started to
report the following warning:

```
40/50 points: code has no errors, warnings, lints, or formatting issues

INFO: 'onPopInvoked' is deprecated and shouldn't be used. Override onPopInvokedWithResult instead. This feature was deprecated after v3.22.0-12.0.pre.

lib/src/modal/modal_sheet.dart:395:13

    ╷
395 │       route.onPopInvoked(didPop);
    │             ^^^^^^^^^^^^
```

This decreases the package's pub points for the *Pass-static-analysis*
section from 50 to 40, but the current CI requires it to always be 50,
resulting in no PR passing the CI.

Although it is possible to fix this warning by following [the migration
guide](https://docs.flutter.dev/release/breaking-changes/popscope-with-result),
we are not doing this for now as it breaks backward compatibility.
Instead, this PR works around this CI issue by decreasing the required
pub points for that section to 40 until Flutter 3.24 is widely adopted
in the community.

## Summary (check all that apply)

- [ ] Modified / added code
- [ ] Modified / added tests
- [ ] Modified / added examples
- [x] Modified / added others (pubspec.yaml, workflows, etc...)
- [ ] Updated README
- [ ] Contains breaking changes
  - [ ] Created / updated migration guide
- [ ] Incremented version number
  - [ ] Updated CHANGELOG