Fix: Press-and-hold gesture in PageView doesn't stop momentum scrolling (#219)
## Fixes / Closes (optional)
<!-- List any issues or pull requests that this PR fixes or closes. Use
the format: "Fixes #123" or "Closes #456". -->

Fixes #214.

## Description

This PR adds `holdWithScrollPosition()` to `ScrollableSheetExtent` to
handle press-and-hold gestures in scrollable sheet content. It starts a
`HoldScrollDrivenSheetActivity`, which stops momentum scrolling (if it
is running).

The absence of this handling logic hasn't been problematic for simple
ListViews or PageViews with only one page. This is because the SDK
triggers `ScrollPosition.drag()` immediately after calling
`ScrollPosition.hold()` when the user performs a press-and-hold gesture.
However, in a PageView with multiple ListViews, `drag()` won't be called
after `hold()` until the user actually moves their finger.

## Summary (check all that apply)
<!-- Mark the boxes that apply to this PR. Add details if necessary. -->
- [x] Modified / added code
- [x] Modified / added tests
- [x] Modified / added examples
- [ ] Modified / added others (pubspec.yaml, workflows, etc...)
- [ ] Updated README
- [ ] Contains breaking changes
  - [ ] Created / updated migration guide
- [x] Incremented version number
  - [x] Updated CHANGELOG