Fix: Modal sheet jerks when swipe to dismiss (#258)
## Related issues (optional)

Fixes #250.

## Description

This PR modifies `_SwipeDismissibleController` to ensure that the
transition animation curve remains linear during the animation triggered
by a swipe-to-dismiss gesture. This is crucial to prevent the sheet from
jerking when the user swipes it down, as reported in #250. See the
comments in the `_handleDragEnd` method for a more detailed explanation.

## Summary (check all that apply)

- [x] Modified / added code
- [x] Modified / added tests
- [ ] Modified / added examples
- [x] Modified / added others (pubspec.yaml, workflows, etc...)
- [ ] Updated README
- [ ] Contains breaking changes
  - [ ] Created / updated migration guide
- [ ] Incremented version number
  - [ ] Updated CHANGELOG