Fix: initial offset of PagedSheet is ignored when using auto_route (#310)
## Problem / Issue

Fixes #308.

The `AutoRouter` widget from the `auto_route` package, which wraps the
`Navigator`, requires at least two frames to build the navigator (see
[this
code](https://github.com/Milad-Akarie/auto_route_library/blob/6feeda78d004d08dfc8fbcb6dde83200fd4edd4c/auto_route/lib/src/router/widgets/auto_router.dart#L145-L149)
for more details). As a result, in the frame where the `PagedSheet` is
first built, the nested navigator does not yet exist. The `PagedSheet`
attempts to use the `initialOffset` property of the initial route from
the nested navigator to determine its position on first build. However,
since the navigator and its route do not exist at that point, the sheet
cannot access the `initialOffset`, and as a result, the initial offset
is effectively ignored.

## Solution

Introduce `_PagedSheetIdleActivity`, a specialized type of
`SheetIdleActivity` for `PagedSheet`. It uses the `initialOffset` of the
current route as the `SheetIdleActivity.targetOffset` if the route has
not been laid out yet. When the current route changes, the `PagedSheet`
model starts a new `_PagedSheetIdleActivity` and if the size of that
route is not determined, its `initialOffset` is used as the target
offset. Then, during the subsequent layout phase, once the route is laid
out, `_PagedSheetIdleActivity.applyNewLayout` is called with the
finalized route size. This updates the sheet's offset to match the
`initialOffset`.

With this change, the `initialOffset` of a route can now be correctly
applied even if the navigator was not available during the initial build
of the `PagedSheet`.