fix: PagedSheet ignores initialOffset when using auto_route and the first page is fullscreen (#321)
## Problem / Issue

Fixes #315

When using `PagedSheet` with `auto_route`, if the initial route rendered
a fullscreen page, the `initialOffset` specified in the
`PagedSheetRoute` was ignored. This happened because `AutoRouter`'s
nested `Navigator` builds one frame after the `PagedSheet` initially
builds (the relevant code is
[here](https://github.com/Milad-Akarie/auto_route_library/blob/6feeda78d004d08dfc8fbcb6dde83200fd4edd4c/auto_route/lib/src/router/widgets/auto_router.dart#L145-L149)),
causing the layout update logic in `_PagedSheetModel` to skip applying
the initial offset in certain conditions.

## Solution

A workaround has been implemented in `_PagedSheetModel.applyNewLayout`.
When the sheet is first built and the nested Navigator hasn't built yet,
it temporarily applies a zero-sized layout. This ensures that in the
next frame, when the Navigator builds and the correct layout is
provided, `applyNewLayout` correctly detects a layout change and applies
the `initialOffset` as expected.