refactor(pkg): Hide SheetDismissible from public API (#343)
## Problem / Issue

`SheetDismissible` was reintroduced in #339. However, while working on
#170, we realized that the fix for #170 isn't compatible with exposing
`SheetDismissible` as part of the public API. More details will be
available in an upcoming PR.

## Solution

Since `SheetDismissible` hasn't been published yet, this PR simply makes
it a private class. We're not removing it entirely, as it will be
required to address other issues like #303. Instead, users are
encouraged to use `PopScope.canPop` to control the dismissibility of a
modal in the build method.