feat(pkg): Add SheetPopScope to enable/disable the swipe gesture in modals from within build method (#359)
## Problem / Issue

- Closes https://github.com/fujidaiti/smooth_sheets/issues/356
- Related to:
  - https://github.com/fujidaiti/smooth_sheets/issues/346
  - https://github.com/fujidaiti/smooth_sheets/issues/249#issuecomment-2822457377

Currently, while `PopScope` can influence the dismiss behavior, the
swipe-to-dismiss gesture for a modal sheet is primarily controlled at
the route level by `ModalSheetRoute.swipeDismissible`. This offers
limited direct means to dynamically enable or disable the gesture from
*within* a widget (i.e., during the `build` method of a widget inside
the sheet), particularly when needing to disable the gesture and prevent
the route from being popped. This limits the flexibility in scenarios
where the ability to dismiss the sheet via a swipe gesture needs to
change based on the internal state of the sheet's content.

## Solution

This PR introduces `SheetPopScope`, a new widget that provides
fine-grained control over the pop behavior and the swipe-to-dismiss
gesture of a modal sheet, directly from within the sheet's widget tree.