feat(pkg): Changed SwipeDismissSensitivity to use a SheetOffset as the minimum drag value (#415)
### Problem

The current `SwipeDismissSensitivity.minDragDistance` only accepts a fixed drag distance in pixels.
However, this is inflexible in some cases. For example, when using a `PagedSheet`, the ideal drag distance
to dismiss the modal sheet may differ for each page.

### Proposal

Remove `SwipeDismissSensitivity.minDragDistance` and add `SwipeDismissSensitivity.dismissalOffset` instead.
It defines the threshold in terms of `SheetOffset`, below which the sheet will be dismissed when the drag ends.
This parameter introduces great flexibility in configuring the threshold, such as making it depend on the content size
or the viewport size:

- With `SwipeDismissSensitivity(dismissalOffset: SheetOffset(0.4))`, the sheet will dismiss if 40% or less
   of the sheet is visible when the drag ends.
- With `SwipeDismissSensitivity(dismissalOffset: SheetOffset.absolute(200))`, the sheet will dismiss if 200 pixels
   or less of the sheet is visible when the drag ends.

Since `SheetOffset` is an interface, we can also define a custom threshold class for more complex use cases:

```dart
SwipeDismissSensitivity(dismissalOffset: const CustomThreshold());

class CustomThreshold implements SheetOffset {
  const CustomThreshold();
  
  @override
  double resolve(ViewportLayout metrics) {
    return max(metrics.contentSize.height * 0.5, 80);
  }
}
```

---------

BREAKING CHANGE: Added `SwipeDismissSensitivity.dismissalOffset` and `SwipeDismissSensitivity.minDragDistance` was removed instead.

Co-authored-by: Bjarte Bore <bjarte.bore@lyse.no>